* sysdeps/mach/hurd/mmap.c (__mmap): Add a dead case that quiets a
	warning.

	* sysdeps/mach/hurd/getcwd.c
	(_hurd_canonicalize_directory_name_internal): unsigned int -> size_t

+++ sysdeps/mach/hurd/getcwd.c
@@ -118,7 +118,7 @@ _hurd_canonicalize_directory_name_internal (file_t thisdir,
       int mount_point;
       file_t newp;
       char *dirdata;
-      unsigned int dirdatasize;
+      size_t dirdatasize;
       int direntry, nentries;
